refactor fuse flags and cmds - mount & flags

+++ pkg/common/config/fuse.go
@@ -48,8 +48,8 @@ var defaultFuseConfig = FuseConfig{
 		UserName:             "root",
 		EntryTimeout:         1,
 		AttrTimeout:          1,
-		Uid:                  uint32(os.Getuid()),
-		Gid:                  uint32(os.Getgid()),
+		Uid:                  os.Getuid(),
+		Gid:                  os.Getgid(),
 		IgnoreSecurityLabels: true,
 		DisableXAttrs:        true,
 		AllowOther:           true,
@@ -94,8 +94,8 @@ type Fuse struct {
 	UserName             string `yaml:"userName"`
 	EntryTimeout         int    `yaml:"entryTimeout"`
 	AttrTimeout          int    `yaml:"attrTimeout"`
-	Uid                  uint32 `yaml:"uid"`
-	Gid                  uint32 `yaml:"gid"`
+	Uid                  int    `yaml:"uid"`
+	Gid                  int    `yaml:"gid"`
 	IgnoreSecurityLabels bool   `yaml:"ignoreSecurityLabels"`
 	DisableXAttrs        bool   `yaml:"disableXAttrs"`
 	AllowOther           bool   `yaml:"allowOther"`
